+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.0.3 - 2022-??-?? - Support the root
 
+* Add internationalized domain name support (IDN)
 * Enable SUPPORTS_ROOT_NS for management of root NS records. Requires
   octodns>=0.9.16.