+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## v0.0.3 - 2022-??-?? - Support the root
+## v0.0.3 - 2023-??-?? - Support the root
 
 * Enable SUPPORTS_ROOT_NS for management of root NS records. Requires
   octodns>=0.9.16.
+* Use HTTP 1.1 in for NS1 dynamic DNS HTTPS monitors
 
 ## v0.0.2 - 2022-02-02 - pycountry-convert install_requires
 
@@ -10,7 +11,7 @@
 
 ## v0.0.1 - 2022-01-03 - Moving
 
-#### Nothworthy Changes
+#### Noteworthy Changes
 
 * Initial extraction of Ns1Provider from octoDNS core